+++ tapo-py/src/handlers.rs
@@ -1,5 +1,5 @@
-mod energy_monitoring_plug_handler;
+mod plug_energy_monitoring_handler;
 mod plug_handler;
 
-pub use energy_monitoring_plug_handler::*;
+pub use plug_energy_monitoring_handler::*;
 pub use plug_handler::*;