+++ tapo-py/src/lib.rs
@@ -3,16 +3,16 @@ mod errors;
 mod handlers;
 
 use pyo3::prelude::*;
-use tapo::responses::PlugDeviceInfoResult;
+use tapo::responses::DeviceInfoPlugResult;
 
 use api_client::PyApiClient;
-use handlers::PyEnergyMonitoringPlugHandler;
+use handlers::PyPlugEnergyMonitoringHandler;
 
 #[pymodule]
 #[pyo3(name = "tapo")]
 fn tapo_py(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<PyApiClient>()?;
-    m.add_class::<PyEnergyMonitoringPlugHandler>()?;
-    m.add_class::<PlugDeviceInfoResult>()?;
+    m.add_class::<PyPlugEnergyMonitoringHandler>()?;
+    m.add_class::<DeviceInfoPlugResult>()?;
     Ok(())
 }