+++ tapo/src/responses/device_info_result.rs
@@ -1,7 +1,7 @@
 mod color_light;
 mod color_light_strip;
 mod default_state;
-mod generic_device;
+mod generic;
 mod hub;
 mod light;
 mod plug;
@@ -9,7 +9,7 @@ mod plug;
 pub use color_light::*;
 pub use color_light_strip::*;
 pub use default_state::*;
-pub use generic_device::*;
+pub use generic::*;
 pub use hub::*;
 pub use light::*;
 pub use plug::*;