+++ tapo/src/responses/device_usage_energy_monitoring_result.rs
@@ -1,19 +1,23 @@
 use serde::{Deserialize, Serialize};
 
-use crate::responses::TapoResponseExt;
+use super::{TapoResponseExt, UsageByPeriodResult};
 
-/// Contains the time usage.
+/// Contains the time usage, the power consumption, and the energy savings of the device.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
-pub struct DeviceUsageResult {
+pub struct DeviceUsageEnergyMonitoringResult {
     /// Time usage in minutes.
     pub time_usage: UsageByPeriodResult,
+    /// Power usage in watt-hour (Wh).
+    pub power_usage: UsageByPeriodResult,
+    /// Saved power in watt-hour (Wh).
+    pub saved_power: UsageByPeriodResult,
 }
-impl TapoResponseExt for DeviceUsageResult {}
+impl TapoResponseExt for DeviceUsageEnergyMonitoringResult {}
 
 #[cfg(feature = "python")]
 #[pyo3::pymethods]
-impl DeviceUsageResult {
+impl DeviceUsageEnergyMonitoringResult {
     /// Get all the properties of this result as a dictionary.
     pub fn to_dict<'a>(&self, py: pyo3::Python<'a>) -> pyo3::PyResult<&'a pyo3::types::PyDict> {
         let serialized = serde_json::to_value(self)
@@ -27,16 +31,4 @@ impl DeviceUsageResult {
             Ok(pyo3::types::PyDict::new(py))
         }
     }
-}
-
-/// Usage by period result for today, the past 7 days, and the past 30 days.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
-pub struct UsageByPeriodResult {
-    /// Today.
-    pub today: u64,
-    /// Past 7 days.
-    pub past7: u64,
-    /// Past 30 days.
-    pub past30: u64,
 }